i think we have new generators

+++ lib/hecks/commands/generate.rb
@@ -13,24 +13,32 @@ class HecksDomain
         File.dirname(__FILE__) + '/../templates'
       end
 
-      desc 'generate_from_domain_file', ''
-      def generate_from_domain_file
-        instance_eval(domain_file).domain.tap do |domain|
-          generate_domain_objects(domain)
-          generate_operations(domain)
+      desc 'new', 'Generate an empty domain project'
+      argument :domain_name, optional: true
+      def new
+        if domain_file
+          generate_domain_objects(domain_get)
+          generate_operations(domain_get)
+          generate_domain_files(domain_get)
+        else
+          Generators::Domain.new([domain_name]).invoke_all
         end
-      end
-
-      desc 'new_domain', ''
-      method_option :name, type: :string
-      def new_domain
-        Generators::Domain.new([options[:name]]).invoke_all
       end
 
       private
 
+      def domain_get
+        instance_eval(domain_file).domain
+      end
+
       def domain_file
+        return unless File.exist?(HECKS_FILE_NAME)
+
         File.open(HECKS_FILE_NAME).read
+      end
+
+      def generate_domain_files(domain)
+        Generators::DomainFiles.new([domain]).invoke_all
       end
 
       def generate_operations(domain)
